refactor : postUpdate() api restdocs 헤더값 추가
